🔨fix: 타입 에러 수정
