+++ src/app/types/commit.ts
@@ -38,26 +38,12 @@ interface CommitDetail {
   commitConclusion: string;
 }
 
-interface RequirementCheck {
-  requirement: string;
-  isSatisfied: boolean;
+interface GithubCommit {
+  commitId: string;
+  author: string;
+  commitDate: string;
+  commitMessage: string;
+  files: CommitFile[];
 }
 
-interface ReportData {
-  reportId: string;
-  reportTitle: string;
-  reportSummary: string;
-  commits: CommitDetail[];
-  reportConclusion: string;
-  requirementsCheck: RequirementCheck[];
-  repositoryUrl: string;
-  branch: string;
-}
-
-export type {
-  CommitFile,
-  Analysis,
-  CommitDetail,
-  RequirementCheck,
-  ReportData,
-};
+export type { CommitFile, Analysis, CommitDetail, GithubCommit };
