--- src/app/types/mock_commitInfo.ts
@@ -1,6 +1,0 @@
-type CommitInfo = {
-  id: string;
-  message: string;
-};
-
-export type { CommitInfo };
